update move db to /boot
